pyproject.toml
[project]
name = "tool-use-chat"
version = "0.1.0"
description = "A modular CLI chatbot with OpenAI integration and tool support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
]

